pyproject.toml
[project]
name = "prueba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.0.0b3",
    "azure-identity>=1.26.0b1",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
]
